refactored capture pieces
