+++ client/src/components/CapturedPieces.js
@@ -1,25 +1,14 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 import './css/CapturedPieces.css';
 
-class CapturedPieces extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-    };
-  }
-
-  componentDidMount() {
-  }
-
-  render() {
-    return (
-      <div className="captured-pieces">
-        <h5>Captured Pieces: {this.props.color}</h5>
-      </div>
-    );
-  }
-
-}
+const CapturedPieces = ({ color, capturedPieces }) => {
+  return (
+    <div className="captured-pieces">
+      <h5>Captured Pieces: {color}</h5>
+      {JSON.stringify(capturedPieces)}
+    </div>
+  );
+};
 
 export default CapturedPieces;
